Add ‘set-font!’ and ‘builtin-font’.
We add on this branch with no immediate intention of merging into
upstream since it turns out the fonts snarfed from SDL_gfx subdir
‘Fonts’ display the charset ‘cp1252’, which is a bad-old-days
originated abomination that is best left to die quietly in
obscurity.

Another sticking point is that the interface to ‘set-font!’ is not
really satisfactory from a future-proofing design pov.  It would be
better to introduce a smob that holds all font-related info, create
another user-visible procedure (‘make-font’, or some such), and make
‘builtin-font’ use it (or an internal variant) to return the smob.

See fnt2c comments for how to generate src/sdlgfx-fonts.c,
which is not included in this commit.

Aside from the design issues, the implementation is still
buggy; it segfaults probably due to incorrect memory handling.